Rename ratings.js as rating_groups.js
