+++ src/components/Map.jsx
@@ -3,7 +3,7 @@ import { icons }                from "vm-leaflet-icons"
 import URLs                     from "../constants/URLs"
 import { IDS }                  from  "../constants/Categories"
 import STYLE                    from "./styling/Variables"
-import { avg_rating_for_entry } from "../rating"
+import { avg_rating_for_entry } from "../rating_groups"
 import styled                   from "styled-components";
 import T                        from "prop-types";
 import { FontAwesomeIcon }      from '@fortawesome/react-fontawesome'
@@ -90,7 +90,7 @@ class KVMMap extends Component {
             avg_rating = avg_rating_for_entry(ratings_for_entry);
           }
 
-          if((e.ratings && e.ratings.length > 0 && avg_rating && avg_rating > 0) 
+          if((e.ratings && e.ratings.length > 0 && avg_rating && avg_rating > 0)
             || e.categories[0] === EVENT){
             let opacity = 0.5;
             if(highlight.indexOf(e.id) == 0 || highlight.length == 0) opacity = 1;
