+++ src/rating_groups.js
